pyproject.toml
[project]
name = "master-control"
version = "0.1.0"
description = "Python-based software orchestrator for services, agents, and scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "structlog>=24.0",
    "aiosqlite>=0.20",
    "croniter>=2.0",
    "rich>=13.0",
]

[project.scripts]
master-control = "master_control.cli.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "httpx>=0.27",
    "psutil>=6.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
